[refactor] :recycle: adapt controller to new types and services (#276)
