+++ src/controllers/ImageController.ts
@@ -1,9 +1,11 @@
 import { FastifyReply, FastifyRequest, RouteHandlerMethod } from 'fastify';
 
 import { Logger } from '../helpers/loggerHelper';
-import { getUser } from '../services/mongo/mongoService';
+import { icpService } from '../services/icp/icpService';
 import { isShortUrl } from '../helpers/validationHelper';
-import { uploadToICP, uploadToIpfs, downloadAndProcessImage } from '../services/uploadService';
+import { ipfsService } from '../services/ipfs/ipfsService';
+import { downloadAndProcessImage } from '../services/imageService';
+import { mongoUserService } from '../services/mongo/mongoUserService';
 import {
   returnErrorResponse,
   returnSuccessResponse,
@@ -33,8 +35,8 @@ async function processAndUploadImage(imageUrl: string, fileName: string): Promis
   try {
     const processedImageBuffer = await downloadAndProcessImage(imageUrl);
 
-    const icpUrl = await uploadToICP(processedImageBuffer, fileName);
-    const ipfsUrl = await uploadToIpfs(processedImageBuffer, fileName);
+    const icpUrl = await icpService.uploadToICP(processedImageBuffer, fileName);
+    const ipfsUrl = await ipfsService.uploadToIpfs(processedImageBuffer, fileName);
 
     return {
       success: true,
@@ -84,7 +86,7 @@ export const uploadImage: RouteHandlerMethod = async (
       return await returnErrorResponse(reply, 400, 'Short Url not allowed');
     }
 
-    const user = await getUser(phone_number);
+    const user = await mongoUserService.getUser(phone_number);
     if (!user) {
       Logger.warn('uploadImage', `User not found: ${phone_number}`);
       return await returnErrorResponse(reply, 404, 'User not found');
