+++ src/controllers/rampController.ts
@@ -1,7 +1,7 @@
 import { FastifyReply, FastifyRequest } from 'fastify';
 
 import { Logger } from '../helpers/loggerHelper';
-import { MantecaUserBalance } from '../types/manteca';
+import { MantecaUserBalance } from '../types/mantecaType';
 import { mantecaUserService } from '../services/manteca/user/mantecaUserService';
 import { mantecaPriceService } from '../services/manteca/market/mantecaPriceService';
 import { returnErrorResponse, returnSuccessResponse } from '../helpers/requestHelper';
@@ -81,7 +81,7 @@ export const createRampUser = async (request: FastifyRequest, reply: FastifyRepl
   };
   */
 
-  // TO-REVIEW: Falla el create User en el sandbox, porque no tiene el campo address y no est· en la documentaciÛn como se completa.
+  // TO-REVIEW: Falla el create User en el sandbox, porque no tiene el campo address y no est√° en la documentaci√≥n como se completa.
   // https://docs.manteca.dev/api-runner/mantecadev/cripto/gestion-de-usuarios/usuarios-1/crear-usuario
   const userCreated =
     // await mantecaUserService.createUser(mockCreateUser);
